\name{distcompare}
\alias{distcompare}
\title{Distribution Comparison}
\description{
Distribution Comparison (DistCompare) is a toolset for visualizing comparison between two distributions.
}
\details{
This toolset contains functions that plot:

1) two distributions side by side for comparison (\code{\link{comp.dist.plot}}), and

2) Cliff's Delta effect size and the confidence interval (\code{\link{effsize.range.plot}}).

For a complete list of functions, use \code{help(package = "distcompare")}
}
\author{
Maintainer: Dayi Lin \url{http://lindayi.me}
}
\references{
Please consider citing the following paper when using this package and the visualization style of effect size:

Lin, D., Bezemer, C. P., & Hassan, A. E. (2018). \href{https://www.researchgate.net/publication/317570653_An_Empirical_Study_of_Early_Access_Games_on_the_Steam_Platform}{An empirical study of early access games on the Steam platform}. Empirical Software Engineering, 23(2), 771-799.

}
